Run REST API tests for 2 server configurations
