+++ tests/python/rest_api/test_tasks.py
@@ -13,7 +13,6 @@ from functools import partial
 from http import HTTPStatus
 from itertools import chain, product
 from math import ceil
-from multiprocessing.pool import ThreadPool
 from pathlib import Path
 from tempfile import NamedTemporaryFile, TemporaryDirectory
 from time import sleep, time
@@ -29,8 +28,6 @@ from cvat_sdk.core.proxies.tasks import ResourceType, Task
 from cvat_sdk.core.uploading import Uploader
 from deepdiff import DeepDiff
 from PIL import Image
-from redis import Redis
-from rq import Queue
 
 import shared.utils.s3 as s3
 from shared.fixtures.init import docker_exec_cvat, kube_exec_cvat
@@ -2687,46 +2684,3 @@ class TestImportWithComplexFilenames:
                 for tes in te.shapes
             ]
         )
-
-
-@pytest.mark.usefixtures("restore_db_per_function")
-@pytest.mark.usefixtures("restore_cvat_data")
-@pytest.mark.usefixtures("restore_redis_db_per_function")
-class TestTaskCreationOrder:
-    _USERNAME = "admin1"
-
-    def _create_task(self, idx: int) -> int:
-        task_spec = {
-            "name": f"Test task {idx}",
-            "labels": [
-                {
-                    "name": "car",
-                }
-            ],
-        }
-
-        task_data = {
-            "image_quality": 75,
-            "client_files": generate_image_files(3),
-        }
-
-        task_id, _ = create_task(self._USERNAME, task_spec, task_data)
-        return task_id
-
-    def test_can_create_tasks_in_well_formed_order(self):
-        RQ_JOB_TEMPLATE = "create:task.id{}"
-        with ThreadPool(5) as pool:
-            task_ids = pool.map(self._create_task, range(10))
-
-        # Check for each task that it has 0 or 1 dependency.
-        # Several tasks should not depend on one task.
-        queue = Queue("import", connection=Redis(port=9379))
-        rq_job_ids = [RQ_JOB_TEMPLATE.format(task_id) for task_id in task_ids]
-        rq_jobs = queue.job_class.fetch_many(rq_job_ids, connection=queue.connection)
-
-        for rq_job in rq_jobs:
-            assert len(rq_job.dependency_ids) <= 1, "Several tasks should not depend on one task"
-
-        assert any(
-            [rq_job.dependency_ids for rq_job in rq_jobs]
-        ), "At least one task should depend on another"
